requirements.txt
python-dotenv==1.0.1
langchain-openai==0.2.2
langchain-core==0.3.10
langchain==0.3.3
langchain-community==0.3.2
beautifulsoup4==4.12.3